feat: add getMetrics endpoint

+++ src/main/java/com/github/devmribeiro/clipply/application/dto/response/DashboardMetricsResponse.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 public record DashboardMetricsResponse(
 		BigDecimal expectedRevenue,
-		Integer totalAppointments,
-		Integer confirmedAppointments,
+		Long totalAppointments,
+		Long confirmedAppointments,
 		Integer occupancyRate,
 		Trend revenueTrend,
 		Trend appointmentsTrend,
